requirements.txt
Django 
django-cors-headers
django-ninja
django-ninja-jwt[crypto]
pydantic[email]
python-decouple
rav
